--- a/src/hyperx/clientserver.nim
+++ b/src/hyperx/clientserver.nim
@@ -394,13 +394,13 @@ proc handshake(client: ClientContext) {.async.} =
       let blobRln = await client.sock.recvInto(addr blob[0], blob.len)
       check blobRln == blob.len, newConnClosedError()
       check blob == preface, newConnError(hyxProtocolError)
-  except QueueClosedError as err:
-    doAssert not client.isConnected
-    debugErr2 err
-    if client.error != nil:
-      raise newError(client.error, err)
-    raise err
-  except OsError, SslError:
+  #except QueueClosedError as err:
+  #  doAssert not client.isConnected
+  #  debugErr2 err
+  #  if client.error != nil:
+  #    raise newError(client.error, err)
+  #  raise err
+  except HyperxError, OsError, SslError:
     let err = getCurrentException()
     debugErr2 err
     # XXX err.msg includes a traceback for SslError but it should not