CI before:

````
finished in 4.08s, 122614.56 req/s, 3.63MB/s
````

CI after:

```
finished in 3.23s, 154794.26 req/s, 4.58MB/s
```