@nitely 2025-03-21
https://github.com/nim-lang/Nim/issues/24024